requirements.txt
# Telegram клиент и бот API
telethon>=1.28.0

# Работа с переменными окружения
python-dotenv>=1.0.0

# Работа с аудио файлами (для adder_mp3.py и ogg_from_mp3.py)
pydub>=0.25.0

# Стандартные библиотеки Python (уже включены в Python)
# asyncio - встроенная библиотека
# datetime - встроенная библиотека  
# getpass - встроенная библиотека
# logging - встроенная библиотека
# os - встроенная библиотека
